Email verification: name the ordered settings save, and keep the policy internal
Three follow-ups from review, none changing behaviour.

The settings save that carries the address is its own factory rather than a flag
on the general one, so a caller opts in by reaching for it rather than by
remembering to pass something.

The dehydrate options move to their own module. They were exported from
query-client only so a test could reach them, which put them in a public
package's API, and they had wedged themselves between getPersistQueryClientPromise
and the docblock describing it.

The rationale for the ordering lives with the shared policy and the hook. The
callers no longer restate it.
